requirements.txt
# Base de données
pymongo==4.6.2
redis==5.0.1

# Sécurité
bcrypt==4.1.2

# Gestion des données
pandas==2.2.0
fpdf==1.7.2
openpyxl==3.1.2

# Utilitaires
python-dotenv==1.0.1

# Interface console
colorama==0.4.6
tabulate==0.9.0

# Validation des données
email-validator==2.1.1

# byson

# bcrypt

# pandas

# fpdf

